+++ cmd/sloth/commands/generate.go
@@ -24,14 +24,15 @@ import (
 )
 
 type generateCommand struct {
-	slosInput            string
-	slosOut              string
-	disableRecordings    bool
-	disableAlerts        bool
-	extraLabels          map[string]string
-	sliPluginsPaths      []string
-	sloPeriodWindowsPath string
-	sloPeriod            string
+	slosInput             string
+	slosOut               string
+	disableRecordings     bool
+	disableAlerts         bool
+	disableOptimizedRules bool
+	extraLabels           map[string]string
+	sliPluginsPaths       []string
+	sloPeriodWindowsPath  string
+	sloPeriod             string
 }
 
 // NewGenerateCommand returns the generate command.
@@ -46,6 +47,7 @@ func NewGenerateCommand(app *kingpin.Application) Command {
 	cmd.Flag("sli-plugins-path", "The path to SLI plugins (can be repeated), if not set it disable plugins support.").Short('p').StringsVar(&c.sliPluginsPaths)
 	cmd.Flag("slo-period-windows-path", "The directory path to custom SLO period windows catalog (replaces default ones).").StringVar(&c.sloPeriodWindowsPath)
 	cmd.Flag("default-slo-period", "The default SLO period windows to be used for the SLOs.").Default("30d").StringVar(&c.sloPeriod)
+	cmd.Flag("disable-optimized-rules", "If enabled it will disable optimized generated rules.").BoolVar(&c.disableOptimizedRules)
 
 	return c
 }
@@ -133,7 +135,7 @@ func (g generateCommand) Run(ctx context.Context, config RootConfig) error {
 				return fmt.Errorf("tried loading raw prometheus SLOs spec, it couldn't: %w", err)
 			}
 
-			err = generatePrometheus(ctx, logger, windowsRepo, g.disableRecordings, g.disableAlerts, g.extraLabels, *slos, out)
+			err = generatePrometheus(ctx, logger, windowsRepo, g.disableRecordings, g.disableAlerts, g.disableOptimizedRules, g.extraLabels, *slos, out)
 			if err != nil {
 				return fmt.Errorf("could not generate Prometheus format rules: %w", err)
 			}
@@ -144,7 +146,7 @@ func (g generateCommand) Run(ctx context.Context, config RootConfig) error {
 				return fmt.Errorf("tried loading Kubernetes prometheus SLOs spec, it couldn't: %w", err)
 			}
 
-			err = generateKubernetes(ctx, logger, windowsRepo, g.disableRecordings, g.disableAlerts, g.extraLabels, *sloGroup, out)
+			err = generateKubernetes(ctx, logger, windowsRepo, g.disableRecordings, g.disableAlerts, g.disableOptimizedRules, g.extraLabels, *sloGroup, out)
 			if err != nil {
 				return fmt.Errorf("could not generate Kubernetes format rules: %w", err)
 			}
@@ -155,7 +157,7 @@ func (g generateCommand) Run(ctx context.Context, config RootConfig) error {
 				return fmt.Errorf("tried loading OpenSLO SLOs spec, it couldn't: %w", err)
 			}
 
-			err = generateOpenSLO(ctx, logger, windowsRepo, g.disableRecordings, g.disableAlerts, g.extraLabels, *slos, out)
+			err = generateOpenSLO(ctx, logger, windowsRepo, g.disableRecordings, g.disableAlerts, g.disableOptimizedRules, g.extraLabels, *slos, out)
 			if err != nil {
 				return fmt.Errorf("could not generate OpenSLO format rules: %w", err)
 			}
@@ -170,7 +172,7 @@ func (g generateCommand) Run(ctx context.Context, config RootConfig) error {
 
 // generatePrometheus generates the SLOs based on a raw regular Prometheus spec format input and
 // outs a Prometheus raw yaml.
-func generatePrometheus(ctx context.Context, logger log.Logger, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts bool, extraLabels map[string]string, slos prometheus.SLOGroup, out io.Writer) error {
+func generatePrometheus(ctx context.Context, logger log.Logger, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts, disableOptimizedRules bool, extraLabels map[string]string, slos prometheus.SLOGroup, out io.Writer) error {
 	logger.Infof("Generating from Prometheus spec")
 	info := info.Info{
 		Version: info.Version,
@@ -178,7 +180,7 @@ func generatePrometheus(ctx context.Context, logger log.Logger, windowsRepo aler
 		Spec:    prometheusv1.Version,
 	}
 
-	result, err := generateRules(ctx, logger, info, windowsRepo, disableRecs, disableAlerts, extraLabels, slos)
+	result, err := generateRules(ctx, logger, info, windowsRepo, disableRecs, disableAlerts, disableOptimizedRules, extraLabels, slos)
 	if err != nil {
 		return err
 	}
@@ -202,7 +204,7 @@ func generatePrometheus(ctx context.Context, logger log.Logger, windowsRepo aler
 
 // generateKubernetes generates the SLOs based on a Kuberentes spec format input and
 // outs a Kubernetes prometheus operator CRD yaml.
-func generateKubernetes(ctx context.Context, logger log.Logger, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts bool, extraLabels map[string]string, sloGroup k8sprometheus.SLOGroup, out io.Writer) error {
+func generateKubernetes(ctx context.Context, logger log.Logger, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts, disableOptimizedRules bool, extraLabels map[string]string, sloGroup k8sprometheus.SLOGroup, out io.Writer) error {
 	logger.Infof("Generating from Kubernetes Prometheus spec")
 
 	info := info.Info{
@@ -210,7 +212,7 @@ func generateKubernetes(ctx context.Context, logger log.Logger, windowsRepo aler
 		Mode:    info.ModeCLIGenKubernetes,
 		Spec:    fmt.Sprintf("%s/%s", kubernetesv1.SchemeGroupVersion.Group, kubernetesv1.SchemeGroupVersion.Version),
 	}
-	result, err := generateRules(ctx, logger, info, windowsRepo, disableRecs, disableAlerts, extraLabels, sloGroup.SLOGroup)
+	result, err := generateRules(ctx, logger, info, windowsRepo, disableRecs, disableAlerts, disableOptimizedRules, extraLabels, sloGroup.SLOGroup)
 	if err != nil {
 		return err
 	}
@@ -234,7 +236,7 @@ func generateKubernetes(ctx context.Context, logger log.Logger, windowsRepo aler
 
 // generateOpenSLO generates the SLOs based on a OpenSLO spec format input and
 // outs a Prometheus raw yaml.
-func generateOpenSLO(ctx context.Context, logger log.Logger, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts bool, extraLabels map[string]string, slos prometheus.SLOGroup, out io.Writer) error {
+func generateOpenSLO(ctx context.Context, logger log.Logger, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts, disableOptimizedRules bool, extraLabels map[string]string, slos prometheus.SLOGroup, out io.Writer) error {
 	logger.Infof("Generating from OpenSLO spec")
 	info := info.Info{
 		Version: info.Version,
@@ -242,7 +244,7 @@ func generateOpenSLO(ctx context.Context, logger log.Logger, windowsRepo alert.W
 		Spec:    openslov1alpha.APIVersion,
 	}
 
-	result, err := generateRules(ctx, logger, info, windowsRepo, disableRecs, disableAlerts, extraLabels, slos)
+	result, err := generateRules(ctx, logger, info, windowsRepo, disableRecs, disableAlerts, disableOptimizedRules, extraLabels, slos)
 	if err != nil {
 		return err
 	}
@@ -266,12 +268,16 @@ func generateOpenSLO(ctx context.Context, logger log.Logger, windowsRepo alert.W
 
 // generate is the main generator logic that all the spec types and storers share. Mainly
 // has the logic of the generate app service.
-func generateRules(ctx context.Context, logger log.Logger, info info.Info, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts bool, extraLabels map[string]string, slos prometheus.SLOGroup) (*generate.Response, error) {
+func generateRules(ctx context.Context, logger log.Logger, info info.Info, windowsRepo alert.WindowsRepo, disableRecs, disableAlerts, disableOptimizedRules bool, extraLabels map[string]string, slos prometheus.SLOGroup) (*generate.Response, error) {
 	// Disable recording rules if required.
 	var sliRuleGen generate.SLIRecordingRulesGenerator = generate.NoopSLIRecordingRulesGenerator
 	var metaRuleGen generate.MetadataRecordingRulesGenerator = generate.NoopMetadataRecordingRulesGenerator
 	if !disableRecs {
-		sliRuleGen = prometheus.SLIRecordingRulesGenerator
+		// Disable optimized rules if required.
+		sliRuleGen = prometheus.OptimizedSLIRecordingRulesGenerator
+		if disableOptimizedRules {
+			sliRuleGen = prometheus.SLIRecordingRulesGenerator
+		}
 		metaRuleGen = prometheus.MetadataRecordingRulesGenerator
 	}